Initial commit of the version 1.0.0 build of my firmware. Built and tested using an original Creality v1.1.5 board with TMC2208 quiet stepper drivers, as well as a genuine BL Touch v3.1 sensor. Currently assumes the probe is using the mount supplied when purchasing the kit from Creality 3D, or a reseller. If using a printed mount, or custom fan shrouds on the hot end, then the offsets may need to be adjusted.
All parts on the test printer are stock with the exception of the upgrade to the v1.1.5 (genuine Creality) board and the addition of the BL Tpuch sensor.

Offsets are in place to center the print on the build plate while using the correct sizing of 220x220x250 for the build area. These offsets are not exact and will be updated in a future revision. However, they are close enough for the moment.

Most of the changes to the default code are based on the tutorials by Teaching Tech on Youtube. However, the movement rates have NOT been updated, rather they still use the defaults in the Ender 3 sample configuration supplied with Marlin. This is simply in an attempt to keep the printer settings as near to stock for initial comparison prints.